requirements.txt
altair
mplfinance
numpy
pandas
pandas-datareader
Pillow
plotly
scikit-learn
scipy
sklearn
streamlit
yfinance